% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_and_install_module}
\alias{check_and_install_module}
\title{Check and install Python module}
\usage{
check_and_install_module(module, .py.venv, ...)
}
\arguments{
\item{module}{a unit-length character string}

\item{.py.venv}{character (optional), specifying path to a Python virtual environment to be used.
Defaults to \code{NULL} in which case the user's/systems default Python environment is used
(see \code{py_config()}).}

\item{...}{Additional parameters passed to \code{\link[reticulate]{py_install}}
or \code{\link[reticulate]{virtualenv_install}} (if \code{.py.venv} is not NULL)}
}
\value{
a logical indicating whether module exists (was or has been installed) after check.
}
\description{
Check and install Python module
}
