% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_laser_language}
\alias{check_laser_language}
\title{Check if input language code(s) is/are supported by LASER}
\usage{
check_laser_language(lang, .supported.langs = laser.languages, .verbose = TRUE)
}
\arguments{
\item{lang}{a character vector specifying language codes (ideally ISO 639-1).}

\item{.supported.langs}{a \code{data.frame} or \code{\link[tibble]{tibble}} with
columns \describe{
    \item{code}{language code used by LASER}
    \item{name}{language name}
    \item{iso2c}{ISO 639-1 language code}
    \item{iso3c}{ISO 639-3 language code}
}}

\item{.verbose}{logical indicating whether or not to raise a warning for unknown/not supported languages.}
}
\value{
a named list of logical. If an element is \code{FALSE}, it has additional attributes
    \describe{
         \item{lang}{inputed language code}
         \item{laser.lang}{logical indicating whether \code{lang} is known to LASER}
         \item{problem}{The warning message raised if executed with \code{.verbose = TRUE}}
    }
}
\description{
Check if input language code(s) is/are supported by LASER
}
\note{
\code{check_laser_language} is vectorized over \code{lang}.
}
\examples{
\dontrun{
  test_langs <- c(
    "de" # German
    , "en" # English
    , "csb" # Kashubian was used to train LASER, but is unknown to ISO 639
    , "unk" # 'unknown' language
  )
  res <- check_laser_language(test_langs)
  unlist(res) # as (named) logical vector
}
}
