% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_laser.R
\name{setup_laser}
\alias{setup_laser}
\title{Setup LASER}
\usage{
setup_laser(
  .py.venv = Sys.getenv("RETICULATE_PYTHON_ENV", NA_character_),
  ...,
  .required.models = c("93langs.fcodes", "93langs.fvocab",
    "bilstm.93langs.2018-12-26.pt")
)
}
\arguments{
\item{.py.venv}{path to a Python virtual environment (see \code{\link[reticulate]{virtualenv_create}})}

\item{...}{additional arguments passed to \code{\link[reticulate]{py_install}}
or \code{\link[reticulate]{virtualenv_install}} (if \code{.py.venv} is not NULL)}

\item{.required.models}{a character vector specifying the model data required to use
the \bold{laserembeddings} Python module.
DO NOT change this argument, unless you know what you're doing.}
}
\value{
logical, indicating whether or not setting up the \bold{laserembeddings} Python module was successful
}
\description{
LASER setup. Interactive function to create a
    Python virtual environment (optional) ,
    installs \bold{laserembeddings} (if not already avaialable),
    and downloads all models.
}
